Fix collecting db credentials when there is no localconf.php yet
On a first deployment to a certain environment, the get-db-config.sh
failed because there is no localconf.php installed yet (that is done
at apply-dataset, which requires the database credentials first).
Now database settings are taken from local/config/.. when
localconf.php does not exist yet. (also works for 6.x and higher)