[FEATURE] Exclude db-table feature in backup now dumps structure of excluded tables
Exclude tables (e.g. caching tables) are not dumped by the backup script.
The new feature does dump the structure in a separate set of sql-files so
you can choose to include the structure when restoring, which will result
in a complete, usable database in which the excluded tables exist but empty.

Restoring those tables is done by just moving the tar.gz file 1 level up
from the directory 'excluded-tables-structure' just before doing the restore.